Feat: DetailViewButton 분리작업
